requirements.txt
appnope==0.1.0
arrow==0.10.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
decorator==4.1.2
enum34==1.1.6
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
lxml==3.8.0
numpy==1.13.1
pandas==0.20.3
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
matplotlib==2.0.2
